+++ frontend/src/pages/EditPage/components/FavoriteTab/index.tsx
@@ -1,6 +1,9 @@
 import * as S from './FavoriteTab.styled';
 import useFavoriteTodoListQuery from '../../hooks/useFavoriteListQuery';
-import TodoEditItem from '../TodoTab/TodoEditItem';
+import TodoEditItem from '../TodoEditItem';
+
+import IconButton from '@/components/Button/IconButton';
+import Icon from '@/components/Icon';
 
 const FavoriteTab = () => {
   const { data: favoriteTodoList } = useFavoriteTodoListQuery();
@@ -10,7 +13,11 @@ const FavoriteTab = () => {
   return (
     <S.FavoriteTabLayout>
       {favoriteTodoList.map((todo) => (
-        <TodoEditItem key={todo.todo_id} todo={todo} />
+        <TodoEditItem
+          key={todo.todo_id}
+          todo={todo}
+          left={<IconButton icon={<Icon icon="PlusCircle" />} />}
+        />
       ))}
     </S.FavoriteTabLayout>
   );
