+++ frontend/src/pages/EditPage/components/RecommendTab/index.tsx
@@ -1,6 +1,9 @@
 import * as S from './RecommendTab.styled';
 import useRecommendTodoListQuery from '../../hooks/useRecommendListQuery';
-import TodoEditItem from '../TodoTab/TodoEditItem';
+import TodoEditItem from '../TodoEditItem';
+
+import IconButton from '@/components/Button/IconButton';
+import Icon from '@/components/Icon';
 
 const RecommendTab = () => {
   const { data: recommendTodoList } = useRecommendTodoListQuery();
@@ -10,7 +13,11 @@ const RecommendTab = () => {
   return (
     <S.RecommendTabLayout>
       {recommendTodoList.map((todo) => (
-        <TodoEditItem key={todo.todo_id} todo={todo} />
+        <TodoEditItem
+          key={todo.todo_id}
+          todo={todo}
+          left={<IconButton icon={<Icon icon="PlusCircle" />} />}
+        />
       ))}
     </S.RecommendTabLayout>
   );
