+++ frontend/src/pages/EditPage/components/TodoTab/index.tsx
@@ -1,4 +1,4 @@
-import TodoEditItem from './TodoEditItem';
+import TodoEditItem from '../TodoEditItem';
 import * as S from './TodoTab.styled';
 
 import IconButton from '@/components/Button/IconButton';
@@ -21,7 +21,12 @@ const TodoTab = () => {
 
       <S.TodoEditList>
         {todoList.map((todo) => (
-          <TodoEditItem key={todo.todo_id} todo={todo} />
+          <TodoEditItem
+            key={todo.todo_id}
+            todo={todo}
+            left={<IconButton icon={<Icon icon="MinusCircle" />} />}
+            right={<IconButton icon={<Icon icon="Edit" />} />}
+          />
         ))}
         <IconButton icon={<Icon icon="PlusCircle" />} text="직접 추가하기" isFull={true} />
       </S.TodoEditList>
